Set background of events to category color (pending team discussion).
Required changing the parameters from MonthItem.select() from boolean to
two displayables. Boolean used to represent whether the two displayables
were the same, so same functionality was preserved. Corresponding calls
to MonthItem.select() throughout code were updated.

This is one of the possible solutions for this. Whether this stays or
not will be discussed in future meeting