Implemented a list of selected categories, month filtering by category.
Implemented a custom listener for check boxes to maintain a list of
selected categories that can be queried to determine which events to
display on calendar. Also implemented calendar filter for month
calendar, so that it only displays events with selected categories. This
may overlap with Zack's work, but had to be implemented to verify that
logic was working as intended. We can sort out any conflicts in today's
meeting.